finance/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
aiosqlite==0.20.0
anthropic==0.42.0
python-dotenv==1.0.1
pydantic==2.10.3
python-jose[cryptography]==3.3.0
httpx==0.28.1
